Move writer to metrics-v3
