+++ lib/datadog-metrics-v3/src/lib.rs
@@ -3,5 +3,15 @@ mod types;
 mod writer;
 
 pub use interner::Interner;
-pub use types::{value_type_for_values, FLAG_HAS_UNIT, FLAG_NO_INDEX, V3MetricType, V3ValueType};
-pub use writer::{V3MetricBuilder, V3Writer};
+pub use types::{value_type_for_values, V3MetricType, V3ValueType, FLAG_HAS_UNIT, FLAG_NO_INDEX};
+pub use writer::{
+    V3ColumnBytes, V3EncodedData, V3EncodedMetrics, V3EncoderStats, V3MetricBuilder, V3ValueEncodingStats, V3Writer,
+    DICT_NAME_STR_FIELD_NUMBER, DICT_ORIGIN_INFO_FIELD_NUMBER, DICT_RESOURCE_LEN_FIELD_NUMBER,
+    DICT_RESOURCE_NAME_FIELD_NUMBER, DICT_RESOURCE_STR_FIELD_NUMBER, DICT_RESOURCE_TYPE_FIELD_NUMBER,
+    DICT_SOURCE_TYPE_NAME_FIELD_NUMBER, DICT_TAGSETS_FIELD_NUMBER, DICT_TAGS_STR_FIELD_NUMBER,
+    DICT_UNIT_STR_FIELD_NUMBER, INTERVALS_FIELD_NUMBER, NAMES_FIELD_NUMBER, NUM_POINTS_FIELD_NUMBER,
+    ORIGIN_INFO_FIELD_NUMBER, RESOURCES_FIELD_NUMBER, SKETCH_BIN_CNTS_FIELD_NUMBER, SKETCH_BIN_KEYS_FIELD_NUMBER,
+    SKETCH_NUM_BINS_FIELD_NUMBER, SOURCE_TYPE_NAME_FIELD_NUMBER, TAGS_FIELD_NUMBER, TIMESTAMPS_FIELD_NUMBER,
+    TYPES_FIELD_NUMBER, UNIT_REFS_FIELD_NUMBER, VALS_FLOAT32_FIELD_NUMBER, VALS_FLOAT64_FIELD_NUMBER,
+    VALS_SINT64_FIELD_NUMBER,
+};
