+++ lib/saluki-components/src/encoders/datadog/metrics/mod.rs
@@ -1572,7 +1572,11 @@ fn write_metric_to_v3(
             resources.insert(device_idx, ("device", device));
         }
     }
-    builder.set_resources(&resources);
+    builder.set_resources(
+        resources
+            .into_iter()
+            .map(|(rtype, rname)| (rtype.to_string(), rname.to_string())),
+    );
 
     // Origin metadata
     if let Some(origin) = metric.metadata().origin() {
@@ -1585,7 +1589,7 @@ fn write_metric_to_v3(
                 subproduct,
                 product_detail,
             } => {
-                builder.set_origin(*product, *subproduct, *product_detail, false);
+                builder.set_origin(*product, *subproduct, *product_detail);
             }
         }
     }
