+++ lib/saluki-components/src/encoders/datadog/metrics/v3/constants.rs
@@ -1,33 +1,7 @@
-// Protocol Buffers field numbers for series and sketch payload messages in the V3 format.
-//
-// These field numbers come from the Protocol Buffers definitions in `lib/protos/datadog/proto/agent-payload/intake_v3.proto`.
-pub const DICT_NAME_STR_FIELD_NUMBER: u32 = 1;
-pub const DICT_TAGS_STR_FIELD_NUMBER: u32 = 2;
-pub const DICT_TAGSETS_FIELD_NUMBER: u32 = 3;
-pub const DICT_RESOURCE_STR_FIELD_NUMBER: u32 = 4;
-pub const DICT_RESOURCE_LEN_FIELD_NUMBER: u32 = 5;
-pub const DICT_RESOURCE_TYPE_FIELD_NUMBER: u32 = 6;
-pub const DICT_RESOURCE_NAME_FIELD_NUMBER: u32 = 7;
-pub const DICT_SOURCE_TYPE_NAME_FIELD_NUMBER: u32 = 8;
-pub const DICT_ORIGIN_INFO_FIELD_NUMBER: u32 = 9;
-pub const TYPES_FIELD_NUMBER: u32 = 10;
-pub const NAMES_FIELD_NUMBER: u32 = 11;
-pub const TAGS_FIELD_NUMBER: u32 = 12;
-pub const RESOURCES_FIELD_NUMBER: u32 = 13;
-pub const INTERVALS_FIELD_NUMBER: u32 = 14;
-pub const NUM_POINTS_FIELD_NUMBER: u32 = 15;
-pub const TIMESTAMPS_FIELD_NUMBER: u32 = 16;
-pub const VALS_SINT64_FIELD_NUMBER: u32 = 17;
-pub const VALS_FLOAT32_FIELD_NUMBER: u32 = 18;
-pub const VALS_FLOAT64_FIELD_NUMBER: u32 = 19;
-pub const SKETCH_NUM_BINS_FIELD_NUMBER: u32 = 20;
-pub const SKETCH_BIN_KEYS_FIELD_NUMBER: u32 = 21;
-pub const SKETCH_BIN_CNTS_FIELD_NUMBER: u32 = 22;
-pub const SOURCE_TYPE_NAME_FIELD_NUMBER: u32 = 23;
-pub const ORIGIN_INFO_FIELD_NUMBER: u32 = 24;
-pub const DICT_UNIT_STR_FIELD_NUMBER: u32 = 25;
-pub const UNIT_REFS_FIELD_NUMBER: u32 = 26;
-
+// Display names for the V3 columns, indexed by their Protocol Buffers field number (from
+// `lib/protos/datadog/proto/agent-payload/intake_v3.proto`, also mirrored in
+// `datadog_metrics_v3::writer` as the `*_FIELD_NUMBER` constants). Index 0 is unused since
+// field numbers start at 1.
 pub(super) const COLUMN_NAMES: [&str; 27] = [
     "reserved",
     "DictNameStr",
