+++ lib/saluki-components/src/encoders/datadog/metrics/v3/mod.rs
@@ -18,10 +18,8 @@
 mod constants;
 mod payload;
 mod telemetry;
-mod writer;
 
+pub(crate) use datadog_metrics_v3::V3EncoderStats;
+pub use datadog_metrics_v3::{V3EncodedMetrics, V3MetricType, V3Writer};
 pub(super) use payload::{V3EncodedRequest, V3PayloadLimits, V3PayloadRequest};
 pub(crate) use telemetry::{V3PayloadSplitReason, V3SerializerTelemetry};
-pub use datadog_metrics_v3::V3MetricType;
-pub(crate) use writer::V3EncoderStats;
-pub use writer::{V3EncodedMetrics, V3Writer};
